Created Index.js
This is the first page of the web-based stock portfolio app.
The design was supposed to feature functions ready to call a database to check if the user's email and password were found in the user table.
However, I was not able to setup the database properly in time.
The code starts by calling the http module in order to create a web server instance for the code to run.