Created registration.js
This is the registration page. Had the database been set up, there would be functions querying the database in order to check if the email addres in the registration form had been used. If the email was valid and not found in the database, the end result would have had the success paragraph appear. Otherwise, the fail paragraph would appear instead. The user table would house email, name, balance, and a salted hashed password.