Create transactions.js
This is the transactions page that would have hosted the orders that the user had placed. The databased housed an orders table with user email as the primary key, which stock they bought, how much it cost, and how many shares were purchased with the order. Functions would have queried the table and populated the below page with the order information.