Created portfolio.js
This is the portfolio page where the stock info is pulled from IEXCloud. I used the iexcloud_api_wrapper module to pull the stock info. Then I draw its information out into an object for storing the symbols and pricing while an array stores the list that will be populated into the webpage. First you can see the object creation.