Added the environment variables
Then I need to change enviroment variables in order for the module to acess IEXCloud. I know it is bad practice to put an API key in the code itself. However, IEXCloud documentation (https://iexcloud.io/docs/api/) states that one of the keys you get are publishable which is the one I am using below.